+++ src/memoize.ts
@@ -1,10 +1,10 @@
-import type { FetchOptions, LRUCacheForClustersAsPromised } from './index.js';
+import type { FetchOptions, LRUCacheClustered } from './index.js';
 
 export type MemoizeOptions = FetchOptions;
 
 // eslint-disable-next-line @typescript-eslint/no-empty-object-type
 export function memoize<Args extends unknown[], K extends {}, V extends {}>(
-  cache: LRUCacheForClustersAsPromised<K, V>,
+  cache: LRUCacheClustered<K, V>,
   fn: (...args: Args) => Promise<V> | V,
   keyFn: (...args: Args) => K,
   opts?: MemoizeOptions,